+++ index.js
@@ -17,6 +17,7 @@ const RateLimiterValkeyGlide = require('./lib/RateLimiterValkeyGlide');
 const RateLimiterSQLite = require('./lib/RateLimiterSQLite');
 const RateLimiterEtcd = require('./lib/RateLimiterEtcd');
 const RateLimiterEtcdNonAtomic = require('./lib/RateLimiterEtcdNonAtomic');
+const RateLimiterQueueError = require('./lib/component/RateLimiterQueueError')
 
 module.exports = {
   RateLimiterRedis,
@@ -40,4 +41,5 @@ module.exports = {
   RateLimiterSQLite,
   RateLimiterEtcd,
   RateLimiterEtcdNonAtomic,
+  RateLimiterQueueError
 };
